Rename GeeData to GeeDataValueSet
