+++ src/domain/gee/entities/GeeDataSet.ts
@@ -1,0 +1,6 @@
+export type GeeDataSet = {
+    displayName: string;
+    pointer: string;
+    bands: number[];
+    doc: number;
+};
